# Refactor RoutePrefix to be a string