@jsumners 2018-01-04
I wonder if the object was there due to V8 classes.