@nwoltman 2018-01-04
Something about keeping the same shape of the fastify instance?