@nwoltman 2018-01-04
Makes the code easier to understand and maintain.