@jsumners 2018-01-04
> Something about keeping the same shape of the fastify instance perhaps?

Exactly. I'd like to hear what @mcollina and/or @delvedor have to say before I add a LGTM.