@nwoltman 2018-01-05
Thanks for the clarification @delvedor :)
I added a new test with multiple `register` calls.